+++ CHANGELOG.md
@@ -4,8 +4,10 @@
   - **JDLImporter**:
     - Previously, it was a class that could be instantiated from either JDL files or string, and a configuration object.
     - Now, in order to create a JDL importer, one must use createImporterFromContent or createImporterFromFiles.
-  - **JDL(entity, field, enum, etc.):**
+  - **JDL(entity, field, enum, etc.)**:
     - Removed validity check methods and moved them to the BusinessErrorChecker or their own validators.
+  - **BinaryOptions**:
+    - Removed the `pager` value from the `pagination` option.
 
 ## What's new
   - Introduced new API, which will be the future API (for v7)